#7.2 pristup zdr. kartonu pomocu razlicitih parametara

+++ spring-boot/src/main/java/com/mrsisa/controller/ClinicController.java
@@ -31,7 +31,7 @@ public class ClinicController {
 	
 	@PreAuthorize("hasAnyRole('PATIENT')")
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET) 
-	public ResponseEntity<Clinic> getOne(@PathVariable Integer id) throws ResourceNotFoundException{		
+	public ResponseEntity<Clinic> getOne(@PathVariable Long id) throws ResourceNotFoundException{		
 		return new ResponseEntity<Clinic>(clinicService.findOne(id), HttpStatus.OK);
 	}
