+++ spring-boot/src/main/java/com/mrsisa/repository/ClinicRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mrsisa.entity.Clinic;
 
-public interface ClinicRepository extends JpaRepository<Clinic, Integer>{
+public interface ClinicRepository extends JpaRepository<Clinic, Long>{
 
 }
