+++ spring-boot/src/main/java/com/mrsisa/repository/UserAccountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mrsisa.entity.UserAccount;
 
-public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
 	Optional<UserAccount> findByEmail(String email);
-	Optional<UserAccount> findById(Integer id);
+	Optional<UserAccount> findById(Long id);
 }
